blog controller updated

+++ blog-jwt-api/app/Http/Controllers/API/BlogController.php
@@ -5,34 +5,55 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use App\Models\Blog;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class BlogController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth:api', ['except' => ['index', 'show']]);
+    }
 
     public function index()
     {
 
-        return Blog::all();
+        return Blog::with('user')->latest()->get();
     }
     public function store(Request $request)
     {
-        $blog = Blog::create($request->all());
-        return $blog;
+
+        $blog = $request->user()->blogs()->create($request->all());
+        return response()->json([
+            'message' => 'Blog başarıyla kaydedildi',
+            'blog' => $blog,
+            'user' => $blog->user,
+        ]);
     }
     public function show(Blog $blog)
     {
-        return $blog;
+        return response()->json([
+            'blog' => $blog,
+            'user' => $blog->user,
+        ]);
     }
     public function update(Request $request, Blog $blog)
     {
+        Gate::authorize('modify', $blog);
+
         $blog->update($request->all());
-        return $blog;
+        return response()->json([
+            'blog' => $blog,
+            'message' => 'Blog güncellendi',
+        ]);
     }
 
     public function destroy(Blog $blog)
     {
+        Gate::authorize('modify', $blog);
 
         $blog->delete();
-        return ["message" => "Blog silindi"];
+        return response()->json([
+            'message' => 'Blog silindi',
+        ]);
     }
 }
